requirement.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.5
pillow==11.2.1
pretty-errors==1.2.25
setuptools==78.1.1
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
